Backend/image_upload/requirements.txt
# Additional requirements for Image Upload feature
# Add these to your main requirements.txt file

# Image processing
Pillow>=10.0.0

# WebSocket support
channels>=4.0.0
channels-redis>=4.1.0
redis>=4.5.0

# QR code generation
qrcode[pil]>=7.4.0

# File storage (optional, for S3)
boto3>=1.26.0
django-storages>=1.13.0

# CORS support
django-cors-headers>=4.0.0

# Additional utilities
python-magic>=0.4.27  # For file type detection
